refactor(cli): move spend/rate meters to tui_detail_meters.go (PRD #1009 M3)
